fix: 실수도 다루기위해 자료형 int에서 double로 수정

+++ src/main/java/calculator/OutputHandler.java
@@ -3,7 +3,7 @@ package calculator;
 public class OutputHandler {
     private static final String RESULT_MESSAGE_PREFIX = "결과 : ";
 
-    public static void printResult(int result) {
+    public static void printResult(double result) {
         System.out.println(RESULT_MESSAGE_PREFIX + result);
     }
 }
